% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePointOperations.R
\name{affineGreyscaleTransformation}
\alias{affineGreyscaleTransformation}
\title{Title}
\usage{
affineGreyscaleTransformation(img, a, b)
}
\arguments{
\item{img}{}

\item{a}{}

\item{b}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
