% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageFilter.R
\name{bilateralFilter}
\alias{bilateralFilter}
\title{Title}
\usage{
bilateralFilter(img, sdSpace, sdTone, patchRange = 5)
}
\arguments{
\item{img}{}

\item{sdSpace}{}

\item{sdTone}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
