% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageFeatures.R
\name{calculateCooccurrence}
\alias{calculateCooccurrence}
\title{Calculate Coocurence}
\usage{
calculateCooccurrence(imageOneChannel, direction = c(1, 1),
  greyvalues = 256)
}
\arguments{
\item{direction}{}

\item{greyvalues}{}
}
\description{
Calculate Coocurence
}
\references{
\insertRef{Haralick1973}{Raspository}
}
