% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageOneChannel.R
\name{calculateFourierSpectrum}
\alias{calculateFourierSpectrum}
\title{Title}
\usage{
calculateFourierSpectrum(img, log.scale = TRUE, shift = TRUE)
}
\arguments{
\item{img}{}

\item{log.scale}{}
}
\description{
Title
}
