% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageMorphologicalFilters.R
\name{circleMask}
\alias{circleMask}
\title{Title}
\usage{
circleMask(radius, method = c("bresenham", "sqrt", "anglefunction"))
}
\arguments{
\item{radius}{}

\item{method}{}
}
\description{
Title
}
\references{
\insertRef{Taubig2014}{Raspository}
}
