% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageConvolutions.R
\name{convolveImage}
\alias{convolveImage}
\title{Title}
\usage{
convolveImage(img, kernel)
}
\arguments{
\item{img}{}

\item{kernel}{}
}
\description{
Title
}
\references{
\insertRef{bookConv}{Raspository}
}
