% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageConvolutions.R
\name{derivateKernel}
\alias{derivateKernel}
\title{Title}
\usage{
derivateKernel(type = "sobel", direction = c("x", "y"))
}
\arguments{
\item{type}{}

\item{direction}{}
}
\description{
Title
}
\references{
\insertRef{bookSobel}{Raspository}
}
