% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetICS.R
\name{diffuseSample}
\alias{diffuseSample}
\title{diffuseSample}
\usage{
diffuseSample(connectivity, networkGenes, S)
}
\arguments{
\item{S}{}
}
\description{
diffuseSample
}
\references{
\insertRef{Dimitrakopoulos2018}{Raspository}

\insertRef{Leiserson2015}{Raspository}

\insertRef{Chung2007}{Raspository}
}
