% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dissipatePixel}
\alias{dissipatePixel}
\title{Calculating the Dissipation of Pixels}
\usage{
dissipatePixel(imgOriginal, minimalTreshold)
}
\value{
the image as matrix after the next time step
}
\description{
Calculating the Dissipation of Pixels
}
\references{
\insertRef{Hagenburg2009}{Raspository}
}
\keyword{internal}
