% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageConversion.R
\name{errorDiffusiondDithering}
\alias{errorDiffusiondDithering}
\title{Title}
\usage{
errorDiffusiondDithering(
  img,
  transformPaletteFunction = round,
  method = c("FS", "mae")
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
\references{
\insertRef{Floyd}{Raspository}

\insertRef{Jarvis1976}{Raspository}
}
