% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fsDithering}
\alias{fsDithering}
\title{Dither the imge with Floyd Steinberg}
\usage{
fsDithering(imgOriginal, transformPaletteFunction)
}
\value{
the image as matrix after dithering
}
\description{
Dither the imge with Floyd Steinberg
}
\references{
\insertRef{Floyd}{Raspository}
}
