% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePointOperations.R
\name{gammaCorrection}
\alias{gammaCorrection}
\title{Title}
\usage{
gammaCorrection(img, gamma)
}
\arguments{
\item{img}{}

\item{gamma}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
