% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convolutions.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Title}
\usage{
gaussianKernel(mean = c(10, 10), sigma = 0.5, n = 20000)
}
\arguments{
\item{mean}{}

\item{sigma}{}

\item{size}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
