% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageBW.R
\docType{class}
\name{imageBW-class}
\alias{imageBW-class}
\alias{imageBW}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{original}}{matrix.}

\item{\code{current}}{matrix.}

\item{\code{operations}}{list.}
}}

