% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageHSV.R
\docType{class}
\name{imageHSV-class}
\alias{imageHSV-class}
\alias{imageHSV}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{original}}{array.}

\item{\code{current}}{array.}

\item{\code{operations}}{list.}
}}

