% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeALS.R
\name{imputeALS}
\alias{imputeALS}
\title{imputeALS}
\usage{
imputeALS(data, lambda = 0.75, r = 10, iters = 80)
}
\arguments{
\item{iters}{}
}
\description{
imputeALS
}
\examples{
library(BEclear)
data("BEclearData")
batchEffect <- calcBatchEffects(
data = ex.data, samples = ex.samples,
adjusted = TRUE, method = "fdr")

mdifs <- batchEffect$med
pvals <- batchEffect$pval

summary <-calcSummary(mdifs, pvals)
cleared.data <- clearBEgenes(ex.data, ex.samples, summary)

result <- imputeALS(cleared.data)

}
\references{
insertRef{Koren2009}{Raspository}

insertRef{Hastie2014}{Raspository}
}
