% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageOneChannel.R
\name{is.imageOneChannel}
\alias{is.imageOneChannel}
\title{Is It A One Channel Image}
\usage{
is.imageOneChannel(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
\code{TRUE}, if the object is of type \code{\link[Raspository]{imageOneChannel}}.
 \code{FALSE} otherwise.
}
\description{
Checks if an object is of type \code{\link[Raspository]{imageOneChannel}}.
}
