% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convolutions.R
\name{laplacianOfGaussian}
\alias{laplacianOfGaussian}
\title{Title}
\usage{
laplacianOfGaussian(img, sigma)
}
\arguments{
\item{img}{}

\item{sigma}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
