% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePointOperations.R
\name{logarithmicDynamicCompression}
\alias{logarithmicDynamicCompression}
\title{Title}
\usage{
logarithmicDynamicCompression(img, c = NULL)
}
\arguments{
\item{img}{}

\item{c}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
