% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meaDithering}
\alias{meaDithering}
\title{Dither the imge with minimized average error}
\usage{
meaDithering(img, transformPaletteFunction)
}
\value{
the image as matrix after dithering
}
\description{
Dither the imge with minimized average error
}
\references{
\insertRef{Jarvis1976}{Raspository}
}
