% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageFeatures.R
\name{min.imageOneChannel}
\alias{min.imageOneChannel}
\title{Title}
\usage{
\method{min}{imageOneChannel}(object)
}
\arguments{
\item{object}{An image of class \code{\link[Raspository]{imageOneChannel}}.}
}
\description{
Title
}
