% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageMorphologicalFilters.R
\name{morphDilate}
\alias{morphDilate}
\title{Title}
\usage{
morphDilate(img, mask = diskMask())
}
\arguments{
\item{img}{}

\item{mask}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
