% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageMorphologicalFilters.R
\name{morphErode}
\alias{morphErode}
\title{Title}
\usage{
morphErode(img, mask = diskMask())
}
\arguments{
\item{img}{}

\item{mask}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
