% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageBW.R
\name{multiplyNormalNoise}
\alias{multiplyNormalNoise}
\title{Title}
\usage{
multiplyNormalNoise(object, sd = NULL)
}
\arguments{
\item{sd}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
