% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetICS.R
\name{netICS}
\alias{netICS}
\title{netICS}
\usage{
netICS(adjacencyMatrix, networkGenes, mutationData, diffExpGenes,
  restartProbability = 0.4)
}
\arguments{
\item{restartProbability}{}
}
\description{
netICS
}
\references{
insertRef{Dimitrakopoulos2018}{Raspository}
}
