% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetICS.R
\name{normaliseAdjacencyMatrix}
\alias{normaliseAdjacencyMatrix}
\title{normaliseAdjacencyMatrix}
\usage{
normaliseAdjacencyMatrix(adjacencyMatrix)
}
\arguments{
\item{adjacencyMatrix}{}
}
\description{
normaliseAdjacencyMatrix
}
\references{
insertRef{Dimitrakopoulos2018}{Raspository}
}
