% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetICS.R
\name{performInsulatedHeatDiffusion}
\alias{performInsulatedHeatDiffusion}
\title{performInsulatedHeatDiffusion}
\usage{
performInsulatedHeatDiffusion(adjacencyMatrix, restartProbability)
}
\arguments{
\item{restartProbability}{}
}
\description{
performInsulatedHeatDiffusion
}
\references{
\insertRef{Dimitrakopoulos2018}{Raspository}
}
