% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageFeatures.R
\name{pixelSum.imageOneChannel}
\alias{pixelSum.imageOneChannel}
\title{Title}
\usage{
pixelSum.imageOneChannel(object)
}
\arguments{
\item{object}{An image of class \code{\link[Raspository]{imageOneChannel}}.}
}
\description{
Title
}
