% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageOperations.R
\name{+.imageOneChannel}
\alias{+.imageOneChannel}
\title{Image Addition}
\usage{
\method{+}{imageOneChannel}(A, B)
}
\arguments{
\item{A}{An image of class \code{\link[Raspository]{imageOneChannel}} or 
a scalar.}

\item{B}{An image of class \code{\link[Raspository]{imageOneChannel}} or 
a scalar.}
}
\value{
The resulting image of class \code{\link[Raspository]{imageOneChannel}}.
}
\description{
Addition of one image with another image or a scalar. At least 
one of the two parameters has to be an image. If both are images, the operation
is executed entrywise.
}
