% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetICS.R
\name{prioritize}
\alias{prioritize}
\title{prioritize}
\usage{
prioritize(
  connectivity,
  connectivityBackward,
  networkGenes,
  mutationData,
  mutationInNetwork,
  diffExpInNetwork
)
}
\arguments{
\item{diffExpInNetwork}{}
}
\description{
prioritize
}
\references{
\insertRef{Dimitrakopoulos2018}{Raspository}
}
