% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageBW.R
\name{PSNR}
\alias{PSNR}
\title{Title}
\usage{
PSNR(object, other = NULL)
}
\arguments{
\item{other}{}
}
\description{
Title
}
\details{
\deqn{PSNR(x, y) = 10 \cdot log_{10}  ( \frac{MAX^2}{MSE(x, y)} )}
\deqn{PSNR(x, y) = 10 \cdot log_{10}(MAX^2) - 10 \cdot log_{10}(MSE(x, y))}
\deqn{PSNR(x, y) = 20 \cdot log_{10}(MAX) - 10 \cdot log_{10}(MSE(x, y))}


\deqn{PSNR(x, y) =  - 10 \cdot log_{10}(MSE(x, y))}
}
\references{
\insertRef{Weickert2019}{Raspository}
}
