% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raspository.R
\docType{package}
\name{Raspository-package}
\alias{Raspository}
\alias{Raspository-package}
\title{The Raspository}
\description{
Raspository-package
}
\author{
\strong{Maintainer}: David Rasp \email{David.J.Rasp@gmail.com} (0000-0003-0164-2163)

}
