% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageNoise.R
\name{saltAndPepperNoise}
\alias{saltAndPepperNoise}
\title{Title}
\usage{
saltAndPepperNoise(object, percentage = 0.2)
}
\arguments{
\item{percentage}{}
}
\description{
Title
}
\references{
\insertRef{Weickert2019}{Raspository}
}
