% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageFeatures.R
\name{skewness.imageOneChannel}
\alias{skewness.imageOneChannel}
\title{Title}
\usage{
skewness.imageOneChannel(object)
}
\arguments{
\item{object}{An image of class \code{\link[Raspository]{imageOneChannel}}.}
}
\description{
Title
}
