% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCA.R
\name{BCA}
\alias{BCA}
\title{Bookmark Coloring Algorithm}
\usage{
BCA(graph, v, retentionCoefficient = 0.4, tol = 0.001)
}
\arguments{
\item{tol}{}
}
\description{
Bookmark Coloring Algorithm
}
\references{
insertRef{Berkhin2006}{Raspository}
}
