% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeALS.R
\name{imputeALS}
\alias{imputeALS}
\title{imputeALS}
\usage{
imputeALS(data, lambda, r, iters, use.biases)
}
\arguments{
\item{data}{a matrix containing missing values that should be imputed}

\item{lambda}{the learning rate}

\item{r}{number of coefficients per sample and feature}

\item{iters}{the number of iterations the ALS should be executed}

\item{use.biases}{TRUE if the bias model should be used}
}
\value{
a matrix containing the original values with missing values replaced
by imputed values
}
\description{
imputeALS
}
\examples{
library(BEclear)
data("BEclearData")
batchEffect <- calcBatchEffects(
data = ex.data, samples = ex.samples,
adjusted = TRUE, method = "fdr")

mdifs <- batchEffect$med
pvals <- batchEffect$pval

summary <-calcSummary(mdifs, pvals)
cleared.data <- clearBEgenes(ex.data, ex.samples, summary)

result <- imputeALS(cleared.data)

}
\references{
\insertRef{Koren2009}{Raspository}

\insertRef{Hastie2014}{Raspository}
}
