% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageConversion.R
\name{lbDithering}
\alias{lbDithering}
\title{Lattice Blotzman Dithering}
\usage{
lbDithering(img, epsilon = 0.5, minimalTreshold = 0.01)
}
\arguments{
\item{img}{}

\item{epsilon}{}

\item{minimalTreshold}{}
}
\description{
Lattice Blotzman Dithering
}
\references{
\insertRef{Hagenburg2009}{Raspository}
}
