% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raspository.R
\docType{package}
\name{Raspository-package}
\alias{Raspository}
\alias{Raspository-package}
\title{The Raspository}
\description{
Raspository-package
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{https://github.com/Livia-Rasp/Raspository/issues}
}

}
\author{
Livia Rasp
}
